Support BLPOP and BRPOP commands in cluster mode
We were incorreclty identifying the keys in the BLPOP and BRPOP Redis
commands in cluster mode. Cluster mode needs to identify keys in
commands to know which shard to send the command too.

To know which arguments of redis commands are redis keys, cluster mode
parses the output of the COMMANDS command, which contains information
about the amount and placement of keys in various commands. One bit of
information is the placement of the last key in a command. For commands
that take an arbitrary amount of keys, Redis documentation specifies the
'last key' value should be -1.

For BLPOP and BRPOP the 'last key' value is -2. I cannot find anything
about this in current Redis documentation, but the meaning of this value
seems to be that the final key is the penultimate argument instead of
the last argument. That fits these commands, which take an arbitrary
amount of keys and then, at the very end, a timeout argument which is
not a key.

This commit changes COMMANDS parsing logic to be support these arbitrary
negative values.